+++ src/pages/AddRecipePage/AddRecipePage.jsx
@@ -2,7 +2,7 @@ import Container from "../../components/Container/Container";
 import PathInfo from "../../components/PathInfo/PathInfo";
 import MainTitle from "../../components/MainTitle/MainTitle";
 import Subtitle from "../../components/Subtitle/Subtitle";
-import AddRecipeForm from "./AddRecipeForm/AddRecipeForm";
+import AddRecipeForm from "../../components/AddRecipeForm/AddRecipeForm";
 
 const AddRecipePage = () => {
   return (
